Introduce folder architecture
