+++ frontend/src/pages/organizer/CreateTournament.tsx
@@ -1,11 +1,11 @@
 import { useState } from "react";
-import "../styles/CreateTournament.css";
-import TennisIcon from "../assets/icons/tennis.svg";
-import PingPongIcon from "../assets/icons/pingpong.svg";
-import BadmintonIcon from "../assets/icons/badminton.svg";
+import "../../styles/CreateTournament.css";
+import TennisIcon from "../../assets/icons/tennis.svg";
+import PingPongIcon from "../../assets/icons/pingpong.svg";
+import BadmintonIcon from "../../assets/icons/badminton.svg";
 import TextField from "@mui/material/TextField";
 import z from "zod";
-import { HTTP_ADDRESS } from '../config.ts';
+import { HTTP_ADDRESS } from '../../config.ts';
 
 const TournamentSchema = z.object({
     type: z.string(),
