+++ frontend/src/pages/register/Register.tsx
@@ -1,4 +1,4 @@
-import "../styles/Auth.css";
+import "../../styles/Auth.css";
 import React, { useState } from "react";
 import { useNavigate } from "react-router-dom";
 import z from "zod";
